printer group view init
