+++ client_src/js/group/drag-and-drop.panel.js
@@ -1,4 +1,4 @@
-import { massDragAndDropId, massDragAndDropStatusId } from "./printer-map.options";
+import { massDragAndDropId, massDragAndDropStatusId } from "./group.options";
 
 export function constructMassDragAndDropPanelHTML() {
   return `
