+++ client_src/js/group/group.worker.js
@@ -1,10 +1,6 @@
 import currentOperations from "../lib/modules/currentOperations";
-import { setPrinterMapState } from "./printer-map.state";
-import { init } from "./printer-map.updater";
-import {
-  monitoringSSEventHandler,
-  monitoringWorkerURL
-} from "../monitoring/monitoring-sse.handler";
+import { setPrinterMapState } from "./group.state";
+import { init } from "./group.updater";
 
 let worker = null;
 let currentView = null;
@@ -24,7 +20,7 @@ export function handleVisibilityChange() {
 
 export function createWebWorker(view) {
   currentView = view;
-  worker = new Worker(monitoringWorkerURL);
+  worker = new Worker("/assets/dist/sse.client.min.js");
   worker.onmessage = async function (event) {
     if (event.data != false) {
       //Update global variables with latest information...
