+++ views/groupView.ejs
@@ -6,7 +6,7 @@
         <div id="printerArea" class="row">
         </div>
     </div>
-    <script src="/assets/dist/printer-map.runner.min.js" type="module"></script>
+    <script src="/assets/dist/group-view.runner.min.js" type="module"></script>
     <script src="/assets/dist/bulk-action-commands.functions.min.js" type="module"></script>
 <% }else{ %>
     <%- include('./addPrinters.ejs') %>
